refactor(frontend): split Layout in components
