+++ frontend/src/index.tsx
@@ -10,9 +10,7 @@ const root = ReactDOM.createRoot(
 );
 root.render(
   <React.StrictMode>
-    {/* <BrowserRouter> */}
     <App />
-    {/* </BrowserRouter> */}
   </React.StrictMode>
 );
