+++ frontend/src/layouts/MainLayout/MainLayout.css
@@ -1,40 +1,3 @@
-.flex-between {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  background-color: #fff !important;
-  position: relative;
-}
-
-.flex-center {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  background-color: #fff !important;
-  height: inherit !important;
-}
-
-.border-bottom::after {
-  content: "";
-  position: absolute;
-  bottom: 0;
-  left: 1em;
-  right: 1em;
-  height: 1px;
-  background-color: rgb(236, 236, 236);
-}
-
 .ant-menu-submenu-title {
   padding-left: 16px !important;
-}
-
-.search {
-  max-width: 500px;
-  width: 60%;
-}
-
-.logo {
-  width: 5em;
 }
