+++ frontend/src/pages/Home/Home.tsx
@@ -1,4 +1,4 @@
-import { Button, Col, Row } from "antd";
+import { Button, Row } from "antd";
 import { useState } from "react";
 import { useNavigate } from "react-router";
 import "./Home.css";
