+++ frontend/src/routes/index.tsx
@@ -3,6 +3,7 @@ import MainLayout from "../layouts/MainLayout/MainLayout";
 import Home from "../pages/Home/Home";
 import Login from "../pages/Login/Login";
 import Signup from "../pages/Signup/Signup";
+import Workspace from "../pages/Workspace/Workspace";
 
 const router = createBrowserRouter([
   {
@@ -23,6 +24,11 @@ const router = createBrowserRouter([
       {
         path: "signup",
         element: <Signup />,
+        // loader: teamLoader,
+      },
+      {
+        path: "workspace",
+        element: <Workspace />,
         // loader: teamLoader,
       },
     ],
